Refactored helper functions into utils folder
